+++ tox.ini
@@ -3,17 +3,20 @@
 
 [tox]
 no_package = True
+skip_missing_interpreters = True
 env_list = lint, unit
 
 [vars]
+application = kafka-benchmark
 src_path = {tox_root}/src
 tests_path = {tox_root}/tests
-# lib_path = {tox_root}/lib/charms/sysbench
+# lib_path = {tox_root}/lib/charms/
 all_path = {[vars]src_path} {[vars]tests_path} # {[vars]lib_path}
 
 [testenv]
 set_env =
     PYTHONPATH = {[vars]src_path}:{tox_root}/lib:{tox_root}/src/benchmark/wrapper
+    PYTHONBREAKPOINT=ipdb.set_trace
     PY_COLORS=1
 allowlist_externals =
     /bin/bash
